requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
PyJWT==2.10.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
